Add code to interface to render plants for each timestep

+++ data_importer/data_importer.h
@@ -59,7 +59,7 @@ namespace data_importer
 				int xs, ys;
 				int specidx;
 				float dbh, height;
-				int nplants;
+                                float nplants;
 			};
 
 			struct filedata
